wip: create local storage
