+++ armadillo/src/main/java/org/molgenis/armadillo/exceptions/StorageException.java
@@ -1,7 +1,11 @@
 package org.molgenis.armadillo.exceptions;
 
 public class StorageException extends RuntimeException {
-  public StorageException(Exception cause) {
-    super(cause);
-  }
+    public StorageException(Exception cause) {
+        super(cause);
+    }
+
+    public StorageException(String cause) {
+        super(cause);
+    }
 }
